repared client
