+++ client/client.py
@@ -1,3 +1,4 @@
+import asyncio
 import base64
 
 import pygame
@@ -17,62 +18,26 @@ def on_remote_close():
     print('Connection closed by server')
 
 
+nw = network.create(HOST_ADDR, on_receive, on_remote_close)
+
+pygame.init()
 screen = pygame.display.set_mode((800, 600))
 pygame.display.set_caption('Client')
 clock = pygame.time.Clock()
-nw = network.create(HOST_ADDR, on_receive, on_remote_close)
 
-
-
-
-
-############################################
-
-# Gérer les événements de l'utilisateur, comme la fermeture de la fenêtre de jeu
-def handle_events():
-    for event in pygame.event.get():
-        if event.type == pygame.QUIT:
-            stop()
-
-# mettre à jour l'état du jeu en fonction des entrées du joueur ou d'autres facteurs
-def update_game():
-    pass
-
-# afficher les éléments graphiques sur l'écran
-def draw():
-    # Effacer l'écran
-    screen.fill((0, 0, 0))  # Remplacez (0, 0, 0) par la couleur d'arrière-plan souhaitée
-
-    # Dessiner les éléments graphiques (sprites, textes, etc.) sur l'écran
-    # Exemple : pygame.draw.rect(screen, (255, 255, 255), pygame.Rect(100, 100, 50, 50))
-
-    # Mettre à jour l'affichage
-    pygame.display.flip()
-
-############################################
-
-
-
-
-
-############################################
-# A priori pas besoin de toucher
-def main_loop():
-    running = True
-    while running:
-        handle_events()  # Gérer les événements de l'utilisateur
-        update_game()   # Mettre à jour l'état du jeu
-        draw()          # Afficher les éléments graphiques sur l'écran
-        clock.tick(60)  # Limiter le framerate à 60 FPS (images par seconde)
-
-############################################
-
-
-def start():
-    pygame.init()
+async def main():
     nw.start()
-
-def stop():
-    pygame.quit()
-    nw.stop()
-    
+    run = True
+    while run:
+        for event in pygame.event.get():
+            if event.type == pygame.QUIT:
+                pygame.quit()
+                nw.stop()
+                return
+            if event.type == pygame.KEYDOWN:
+                keys = pygame.key.get_pressed()
+                if keys[pygame.K_SPACE]:
+                    await nw.send(bytes(keys))
+        pygame.display.update()
+        await asyncio.sleep(0)
+        clock.tick(60)
